Move roles guard into auth
